refactor: move ct meter tests to seperate test file.
